v0.06
__HUGE CHANGES__
Overall changes - Added multi-guild support and creation of users in DB if they don't exist on command usage

Leaderboard - Now posts to an embed with 10 users per page
Register - Added Steam API support and an embed that shows your profile info to confirm its the right one
Sell - Now destroys the item row if item.amount is 0 so /shop doesn't show lots of 0 rows
Setup - Pulls webhook info (channel and ID), staff role and shop channel info and stores it for later use. Also allows custom currency icon to be entered early on (optional)
Setupshop - New command that allows the addition or edit of items in the guild's shop, as well as setting the currency if you wish to change it
Shop - Now posts the shop to an embed for cleaner looks

guildCreate event - Creates the guild in the DB (if not already there) and sends a message to the inviter
guildDelete event - Removes the guild from the DB
ready event - Just posts total users in my DB and how many people are playing the game

TODO:
Add page buttons for leaderboard command
Shop item remove subcommand
Limiting setupshop to admins only?
Either add support for other formats or remind staff that the logon/off format needs to be HH:MM
Also remove users with guild ID from UserItems and Users DB on guild deletion